requirements.txt
langchain
openai
python-dotenv
tiktoken
tqdm
pandas
